Code: Small improvements on outfit system
Changes:
* Use gfxId var to a u16 in SetPlayerAvatarAnimation, and return
  with u16 for GetPlayerAvatarStateTransitionByGraphicsId
* Use u16 for gPlayerAvatarAnimGfxIds
* Use GENDER_COUNT define for the outfit arrays
* Change the i var in GetPlayerAvatarStateTransitionByGraphicsId to
  a u32

+++ src/data/outfit_tables.h
@@ -31,7 +31,7 @@ const u16 gPlayerAvatarGfxIds[OUTFIT_COUNT][4][GENDER_COUNT] =
     }
 };
 
-const u8 gPlayerAvatarAnimGfxIds[OUTFIT_COUNT][3][GENDER_COUNT] =
+const u16 gPlayerAvatarAnimGfxIds[OUTFIT_COUNT][3][GENDER_COUNT] =
 {
     [OUTFIT_A] =
     {
